Change the way files organized
It is going to become hard to maintain, name and understand
what is what and where, and, especially, manage site's links,
if the project has 'flat', 'linear' structure.

Now the idea is we always address to index.php, but, depending on
request's parameters different models and views are included.

Video tutorial:
https://youtu.be/PYB0mNMj5Lk