Style Change in footer, header and login
The footer was modified to eliminate from class pt duplication, and md is not working so I removed it.

In header I include a hyperlink to home in Company Name title and in our cars.

Login container class was modified to center the text and input password class was modified to center "forgot password?".