Create Manager Panel and access control to it
Administrators have just one additional functionality (registered users
management) in comparison to staff, that is why a decision to simply
combine their panels was made.

Because of difference in functionality, 'Users' tab on Management Panel
is not shown to staff without Administrator authorities.

To prevent access to Manager Panel for customers and not authorized
user, the header is populated dynamically by its model,
which also has been added.