sign Up validation client-side and server-side complete
In signupValidation.php, we created validation to the differents inputs in form to avoid wrong by the user if they manipulate the client-side validation.
In styles.css, we created our customized boxes for the server validation response because we need it.
In signup.js, we created all the functionalities for the client-side and server-side validation for the different elements of the form because we need it.
In db.php, we created functions to connect with the db because we need to validate fields in the form and insert data to the db and to avoid to repeat ourself.
In signup.php, we add a php code to avoid admin to create a new user and we created the link for ajax.js
In ajax.js, we created a generic funtion to avoid to repeat ourself.